feat: split out some logic
